# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "Description of your API",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/sendMessageToSubscribers": {
            "post": {
                "tags": [
                    "Messaging"
                ],
                "summary": "Send messages to subscribers",
                "description": "Send messages to subscribers.",
                "operationId": "sendMessageToSubscribers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "message"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "The ID of the user to send the message to.",
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "message": {
                                        "description": "The message to be sent to subscribers.",
                                        "type": "string",
                                        "example": "Hello world!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message sent to subscriber successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/subscribe/chatbot": {
            "post": {
                "tags": [
                    "Chat Bot"
                ],
                "summary": "Subscribe users to a chat bot",
                "description": "Subscribe users to a chat bot.",
                "operationId": "subscribeToChatBot",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userId"
                                ],
                                "properties": {
                                    "userId": {
                                        "description": "The ID of the user to subscribe to the chat bot.",
                                        "type": "integer",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/subscribe/channel": {
            "post": {
                "tags": [
                    "Channel or Chat"
                ],
                "summary": "Subscribe users to a channel or chat",
                "description": "Subscribe users to a channel or chat.",
                "operationId": "subscribeToChannel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userId",
                                    "channelListId"
                                ],
                                "properties": {
                                    "userId": {
                                        "description": "The ID of the user to subscribe.",
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "channelListId": {
                                        "description": "The ID of the channel or chat to subscribe to.",
                                        "type": "integer",
                                        "example": "456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/twitter/webhook": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook to receive responses from messenger API",
                "description": "Webhook to receive responses from messenger API.",
                "operationId": "twitterWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "crc_token"
                                ],
                                "properties": {
                                    "crc_token": {
                                        "description": "The CRC token value.",
                                        "type": "string",
                                        "example": "crc_token_value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}